fix: Use picoinit to manage a second tail command for bringing the logs to STDOUT.
Given the way gitolite executes directly within a ssh session and ends
up writing logs back to the caller, this messes the client output
terribly. In addition, we don't have the logs in docker logs.

Instead write all logs to a separate fixed file and tail the logs from
there.